init: add to GH

+++ storage/postgre.ts
@@ -1,57 +1,58 @@
-import * as initEnv from 'dotenv';
-initEnv.config();
-import * as knexInit from 'knex';
-import ItemType from '../types/item';
+
+import { ItemType } from '../types/item';
+import { knex } from '../db/pg';
 
 
-const {PG_PASSWORD, PG_USERNAME, PG_HOST, PG_DBNAME} = process.env;
-
-const url = process.env.DATABASE_URL
-	?? `postgres://${PG_USERNAME}:${PG_PASSWORD}@${PG_HOST}/${PG_DBNAME}`;
-
-const dbName = PG_DBNAME;
 const tableName = 'todoist';
 
-const knex = knexInit({
-  client: 'pg',
-	connection: url,
-	debug: true,
-});
-
-const listAll = async () => {
+const listAll = async (userId: string) => {
+	if (!userId) {
+		throw new Error('User id must be provided')
+	}
 	return knex(tableName)
-		.select();
+		.select()
+		.where({ userId })
 }
-const getById = async (id: string) => {
+
+const getById = async (userId: string, id: string) => {
+	if (!userId) {
+		throw new Error('User id must be provided')
+	}
 	const list = await knex(tableName)
 		.select()
-		.where({ id });
+		.where({ id, userId });
 
 	return list[0];
 }
-const create = async (item: ItemType) => {
-	const {id, data} = item;
+
+const create = async (userId: string, item: ItemType) => {
+	if (!userId) {
+		throw new Error('User id must be provided')
+	}
+	const {id, title, completed, date} = item;
 	const list = await knex(tableName)
-		.insert({id, data})
+		.insert({id, title, completed, userId, date})
 		.returning("*");
 
 	return list[0];
 }
-const update = async (item: ItemType) => {
-	const {id, data} = item;
+const update = async (userId: string, item: ItemType) => {
+	const {id, title, completed, date} = item;
 	const list = await knex(tableName)
-		.update({id, data})
-		.where({ id })
+		.update({id, title, completed, date})
+		.where({ id, userId })
 		.returning("*");
 
 	return list[0];
 }
-const remove = async (id: string) => {
+const remove = async (userId: string, id: string) => {
+	if (!userId) {
+		throw new Error('User id must be provided')
+	}
 	if (!id) return;
-
 	await knex(tableName)
 		.delete()
-		.where({ id });
+		.where({ id, userId });
 }
